64 bit compliance

+++ Server/characters.c
@@ -98,10 +98,11 @@ int char_age(int connection){
 
     /** public function - see header */
 
-    int age=(clients.client[connection].time_of_last_minute - clients.client[connection].char_created) / (60*60*24);
+    int age=( (int) clients.client[connection].time_of_last_minute - (int) clients.client[connection].char_created) / (60*60*24);
 
     return age;
 }
+
 
 void send_char_details(int connection, const char *char_name){
